+++ src/remark/components/slide-number/slide-number.js
@@ -13,15 +13,28 @@ function SlideNumberViewModel (slide, slideshow) {
 
 function formatSlideNumber (slide, slideshow) {
   var format = slideshow.getSlideNumberFormat()
-    , current = slide.number
-    , total = slideshow.getSlides().length
+    , slides = slideshow.getSlides()
+    , current = getSlideNo(slide, slideshow)
+    , total = getSlideNo(slides[slides.length - 1], slideshow)
     ;
 
   if (typeof format === 'function') {
-    return format(current, total);
+    return format.call(slideshow, current, total);
   }
 
   return format
       .replace('%current%', current)
       .replace('%total%', total);
+}
+
+function getSlideNo (slide, slideshow) {
+  var slides = slideshow.getSlides(), i, slideNo = 0;
+
+  for (i = 0; i <= slide.getSlideIndex() && i < slides.length; ++i) {
+    if (slides[i].properties.count !== 'false') {
+      slideNo += 1;
+    }
+  }
+
+  return Math.max(1, slideNo);
 }