+++ src/remark/controllers/inputs/mouse.js
@@ -9,25 +9,38 @@ exports.unregister = function (events) {
 function addMouseEventListeners (events, options) {
   if (options.click) {
     events.on('click', function (event) {
-      if (event.button === 0) {
+      if (event.target.nodeName === 'A') {
+        // Don't interfere when clicking link
+        return;
+      }
+      else if (event.button === 0) {
         events.emit('gotoNextSlide');
       }
     });
     events.on('contextmenu', function (event) {
+      if (event.target.nodeName === 'A') {
+        // Don't interfere when right-clicking link
+        return;
+      }
       event.preventDefault();
       events.emit('gotoPreviousSlide');
     });
   }
 
   if (options.scroll !== false) {
-    events.on('mousewheel', function (event) {
-      if (event.wheelDeltaY > 0) {
+    var scrollHandler = function (event) {
+      if (event.wheelDeltaY > 0 || event.detail < 0) {
         events.emit('gotoPreviousSlide');
       }
-      else if (event.wheelDeltaY < 0) {
+      else if (event.wheelDeltaY < 0 || event.detail > 0) {
         events.emit('gotoNextSlide');
       }
-    });
+    };
+
+    // IE9, Chrome, Safari, Opera
+    events.on('mousewheel', scrollHandler);
+    // Firefox
+    events.on('DOMMouseScroll', scrollHandler);
   }
 }