+++ src/remark/models/slideshow.js
@@ -3,6 +3,7 @@ var Navigation = require('./slideshow/navigation')
   , utils = require('../utils')
   , Slide = require('./slide')
   , Parser = require('../parser')
+  , macros = require('../macros')
   ;
 
 module.exports = Slideshow;
@@ -29,6 +30,7 @@ function Slideshow (events, options) {
   self.togglePresenterMode = togglePresenterMode;
   self.toggleHelp = toggleHelp;
   self.toggleBlackout = toggleBlackout;
+  self.toggleMirrored = toggleMirrored;
   self.toggleFullscreen = toggleFullscreen;
   self.createClone = createClone;
 
@@ -50,7 +52,7 @@ function Slideshow (events, options) {
   function loadFromString (source) {
     source = source || '';
 
-    slides = createSlides(source);
+    slides = createSlides(source, options);
     expandVariables(slides);
 
     links = {};
@@ -97,6 +99,10 @@ function Slideshow (events, options) {
     events.emit('toggleBlackout');
   }
 
+  function toggleMirrored() {
+    events.emit('toggleMirrored');
+  }
+
   function toggleFullscreen () {
     events.emit('toggleFullscreen');
   }
@@ -120,9 +126,9 @@ function Slideshow (events, options) {
   }
 }
 
-function createSlides (slideshowSource) {
+function createSlides (slideshowSource, options) {
   var parser = new Parser()
-   ,  parsedSlides = parser.parse(slideshowSource)
+   ,  parsedSlides = parser.parse(slideshowSource, macros)
     , slides = []
     , byName = {}
     , layoutSlide
@@ -146,7 +152,13 @@ function createSlides (slideshowSource) {
       template = layoutSlide;
     }
 
-    slideViewModel = new Slide(slides.length + 1, slide, template);
+    if (slide.properties.continued === 'true' &&
+        options.countIncrementalSlides === false &&
+        slide.properties.count === undefined) {
+      slide.properties.count = 'false';
+    }
+
+    slideViewModel = new Slide(slides.length, slide, template);
 
     if (slide.properties.layout === 'true') {
       layoutSlide = slideViewModel;