+++ src/remark/parser.js
@@ -35,14 +35,17 @@ function Parser () { }
  *    ...
  *  ]
  */
-Parser.prototype.parse = function (src) {
-  var lexer = new Lexer(),
+Parser.prototype.parse = function (src, macros) {
+  var self = this,
+      lexer = new Lexer(),
       tokens = lexer.lex(cleanInput(src)),
       slides = [],
 
       // The last item on the stack contains the current slide or
       // content class we're currently appending content to.
       stack = [createSlide()];
+
+  macros = macros || {};
 
   tokens.forEach(function (token) {
     switch (token.type) {
@@ -61,6 +64,24 @@ Parser.prototype.parse = function (src) {
           href: token.href,
           title: token.title
         };
+        break;
+      case 'macro':
+        // Macro
+        var macro = macros[token.name];
+        if (typeof macro !== 'function') {
+          throw new Error('Macro "' + token.name + '" not found. ' +
+              'You need to define macro using remark.macros[\'' +
+              token.name + '\'] = function () { ... };');
+        }
+        var value = macro.apply(token.obj, token.args);
+        if (typeof value === 'string') {
+          value = self.parse(value, macros);
+          appendTo(stack[stack.length - 1], value[0].content[0]);
+        }
+        else {
+          appendTo(stack[stack.length - 1], value === undefined ?
+              '' : value.toString());
+        }
         break;
       case 'content_start':
         // Entering content class, so create stack entry for appending